Static Parameters

+++ Parameters.h
@@ -8,7 +8,7 @@ struct Parameters {
 	static Wave* wave_2 ;
   static float ampl_1;
 	static float ampl_2;
-	static uint32_t period_in_ticks;
+	static uint64_t period_in_ticks;
 
 //	static float freq_1;
 //	static float freq_2;
